+++ python/guiderActor/GuiderState.py
@@ -363,17 +363,31 @@ class GuiderState(object):
         else:
             raise RuntimeError('Unknown guide mode %s' % mode)
 
-    def setRefractionBalance(self, plateType, surveyMode):
+    def setRefractionBalance(self, plateType, surveyMode, cmd=None):
         """Set the refraction balance based on the survey name."""
+
+        self.refractionBalance = 1
+
         if plateType == 'APOGEE' or plateType == 'APOGEE-2':
-            self.refractionBalance = 1
+            self.guideWavelength = 5400
         elif (plateType == 'APOGEE&MaNGA' or plateType == 'APOGEE-2&MaNGA') and \
              (surveyMode == 'APOGEE lead'):
-            self.refractionBalance = 1
-        elif plateType == 'BHM&MWM' and surveyMode == 'MWM lead':
-            self.refractionBalance = 1
-        else:
+            self.guideWavelength = 16600
+        elif plateType == 'BHM&MWM':
+            if surveyMode == 'MWM lead':
+                self.guideWavelength = 16600
+            else:
+                self.guideWavelength = 5400
+
+        available_wavelengths = set([wv for gp in self.gprobes.values()
+                                     for wv in gp.haXOffsets])
+        if self.guideWavelength not in available_wavelengths:
             self.refractionBalance = 0
+            if cmd:
+                cmd.warn('text="no guide offsets found for {}A. '
+                         'Offsets will not be applied. '
+                         'Setting refractionBalance=0"'
+                         .format(self.guideWavelength))
 
     def setDecenter(self, decenters, cmd, enable):
         """